requirements.txt
aiofiles==22.1.0
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.1
boto3==1.24.92
botocore==1.27.92
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cssselect==1.1.0
cssutils==2.6.0
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.0rc9
fastapi==0.85.1
ffmpeg-python==0.2.0
future==0.18.2
greenlet==1.1.3.post0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
jmespath==1.0.1
jwt==1.3.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
premailer==3.10.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.28.1
s3transfer==0.6.0
sanic==22.9.0
sanic-routing==22.8.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
websockets==10.3
yagmail==0.15.293
